feat: controllers
- Added a form controller.
- Added an error controller.
- Removing old code from server controller.
- Updated widgets with new controllers
