+++ lib/core/server_binding.dart
@@ -1,5 +1,7 @@
 import 'package:get/get.dart';
+import 'package:obs_for_sama/core/controllers/auth_obs_form_controller.dart';
 import 'package:obs_for_sama/core/controllers/cache_controller.dart';
+import 'package:obs_for_sama/core/controllers/error_controller.dart';
 import 'package:obs_for_sama/core/controllers/scenes_controller.dart';
 import 'package:obs_for_sama/core/controllers/server_controller.dart';
 import 'package:obs_for_sama/core/controllers/sound_controller.dart';
@@ -13,6 +15,8 @@ class ServerBinding extends Bindings {
       ..put<CacheController>(CacheController())
       ..put<ScenesController>(ScenesController())
       ..put<SourcesController>(SourcesController())
-      ..put<SoundController>(SoundController());
+      ..put<SoundController>(SoundController())
+      ..put<AuthObsFormController>(AuthObsFormController())
+      ..put<ErrorController>(ErrorController());
   }
 }
